feat: adição de método em cervice mais views para retornar os produtos que melhor se enquadram no valor target ou meta de pedido

+++ sistem/TPV2/Order/dto.py
@@ -1,4 +1,4 @@
 class OrderItemDTO:
-    def __init__(self, amount: float=0.0, code_product: str=''):
+    def __init__(self, amount: int=0, code_product: str=''):
         self.amount = amount
         self.code_product =  code_product
